+++ cookbooks/arcgis-enterprise/recipes/datastore.rb
@@ -23,6 +23,7 @@ arcgis_enterprise_datastore 'Configure ArcGIS Data Store' do
   install_dir node['arcgis']['data_store']['install_dir']
   data_dir node['arcgis']['data_store']['data_dir']
   types node['arcgis']['data_store']['types']
+  mode node['arcgis']['data_store']['mode']
   run_as_user node['arcgis']['run_as_user']
   server_url 'https://' + node['arcgis']['server']['domain_name'] + ':6443/arcgis'
   username node['arcgis']['server']['admin_username']