+++ cookbooks/esri-tomcat/metadata.rb
@@ -4,11 +4,10 @@ maintainer_email 'contracts@esri.com'
 license          'Apache 2.0'
 description      'Installs/Configures esri-tomcat'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.1.6'
+version          '0.1.7'
 chef_version     '>= 13'
 
 depends          'tomcat', '>= 3.2.0'
-depends          'java', '~> 4.0'
 depends          'openssl', '~> 8.5'
 
 %w(ubuntu redhat centos).each do |os|