+++ cookbooks/arcgis-notebooks/CHANGELOG.md
@@ -2,6 +2,19 @@
 
 This file is used to list changes made in each version of the arcgis-notebooks cookbook.
 
+<<<<<<< HEAD
 # 3.5.0
 
 Initial release of the cookbook.
+=======
+# 3.6.0
+-------
+
+- Added support for ArcGIS Notebook Server 10.8.1.
+- Added unregister_machine recipe.
+
+# 3.5.0
+-------
+
+- Initial release of the cookbook.
+>>>>>>> db62b7ad130ebfa991cabda65f73cf61bc4a7a26